refactor tree methods
